requirements.txt
PyYAML~=6.0.2
playwright~=1.51.0